melp/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-extensions==3.2.1
django-filter==22.1
django-rest==0.8.7
djangorestframework==3.14.0
pytz==2022.5
six==1.16.0
sqlparse==0.4.3
tzdata==2022.5